directory updates and updates to plotinfo clases
