+++ requirements.txt
@@ -4,5 +4,5 @@ tox
 plotly
 dash
 backtrader
-matplotlib
+dacite
 -e .
